prelude: split tracing component code into modules.
